Update on RunningOnRemoteMachine.md
The previous RunningOnRemoteMachine.md is misguiding users on where the WinAppDriver needs to be up and running and to which IP it needs to be pointing at, as well as the differing IP address on the example suggests that a different IP address should be used. This fix should improve understanding.